bnn_competition/requirements.txt
tqdm==4.65.0
torch==1.13.0
torchvision==0.14.0
torchmetrics==0.10.1
pytorch-lightning==1.7.0
torchprofile==0.0.4
imageio==2.22.4
pyyaml==6.0
rich==12.6.0
nip-config==0.8.3
prettytable==3.5.0
